assignment 1 commit
File structure was completely messed up. I need to figure out how GitHub does things when i create a repo, because it seems to create a new folder and copies all my files inside that new folder, which is inside the folder I already created for this project. Note to self *always be aware of what file you're working in*